Call gtk.gdk.threads_init() before doing any GTK+ stuff.
When using the Composite Sampler plugin[1], the UI would freeze (but
the audio would be functional).  The plugin spawns two worker threads
using QtCore's QThread.  Internally, these threads would set up a GLib
event loop for the thread.  In doing so, they would execute:

    if(!g_thread_supported())
        g_thread_init(NULL);

This is according to the GLib docs.[3][4] However, nothing in zynjacku
was causing g_thread_init(NULL) to be called before this, so
g_thread_supported() would return false.  However, some parts of the
GLib main event loop cached some threading info anyway.  When QThread
caused this to be called, it overwrote that cached information and
caused a deadlock on the UI.[7] So, the root problem is that zynjacku
was using GLib's event loop in a threading environment (via
GTK+)... but never caused g_thread_init(NULL) to be called beforehand.

Fast forwarding to the PyGtk module, it says this in the NEWS file:

    Threading support is compiled in by default now.  It is
    necessary to call gtk.threads_init() in order to make use of
    threading though.  There have been a number of fixes to the
    threading support too.
    - pygtk-1.99.17, 31-July-2003[5]

This has since been replaced with gtk.gdk.threads_init()... and is
buried deep in the PyGtk docs.[6]

This patch resolves the issue and also prevents any future such issues
with modules that use GLib.[8]

[1] http://gabe.is-a-geek.org/composite/, version 0.006
[2] http://qt.gitorious.org/qt/qt/blobs/4.7/src/corelib/kernel/qeventdispatcher_glib.cpp
[3] http://library.gnome.org/devel/glib/stable/glib-Threads.html
[4] The 'unstable' glib docs don't mention this procedure, so the
    next release might do this automatically.
[5] http://www.mail-archive.com/pygtk@daa.com.au/msg06241.html
[6] http://library.gnome.org/devel/pygtk/stable/gdk-functions.html#function-gdk--threads-init
[7] The thread with QThread accidentally became the owner of the
    main event loop.  So, when the main event loop when to execute,
    it saw that it wasn't the "owner" and set a wait condition
    to wait for the QThread to release the event loop... which
    was never going to happen.
[8] I spent about 12 hours debugging this.  That's why I'm writing
    a book for a 1-line code change!  :-)